Revert "Merge pull request #393 from bdunne/revert_278"
This reverts commit d27d2219c1a6a75e96519697800dcab057b93581, reversing
changes made to 124c2dd57710780302440f481d10b316e54a6b4c.

+++ lib/manageiq/rpm_build/generate_core.rb
@@ -60,7 +60,7 @@ module ManageIQ
           shell_cmd("yarn install")
           shell_cmd("yarn run available-languages")
           shell_cmd("yarn run build")
-          shell_cmd("git clean -xdf")  # cleanup temp files
+          shell_cmd("git clean -xdf") unless RUBY_PLATFORM.include?("s390x") # git clean would unset the yarn version change
         end
       end
